SE_PROJECT_EXPRESS.V4

+++ controllers/clothingItem.js
@@ -1,5 +1,6 @@
 const ClothingItem = require("../models/clothingItem");
 const {
+  UNAUTH_USER,
   BAD_REQUEST,
   NOT_FOUND,
   DEFAULT,
@@ -69,24 +70,49 @@ const likeItem = (req, res) => {
 
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
+  const owner = req.user._id; // Ensure this is correctly set by middleware
 
-  ClothingItem.findByIdAndDelete(itemId)
-    .orFail()
+  ClothingItem.findById(itemId)
     .then((item) => {
-      res.status(OKAY_STATUS).send({ data: item });
+      if (!item) {
+        return res.status(NOT_FOUND).send({ message: "Item not found." });
+      }
+      if (item.owner.toString() !== owner.toString()) {
+        return res
+          .status(UNAUTH_USER)
+          .send({ message: "You are not authorized to delete this item" });
+      }
+      return ClothingItem.findByIdAndDelete(itemId);
     })
+    .then((deletedItem) => res.send(deletedItem))
     .catch((e) => {
       if (e.name === "CastError") {
         return res.status(BAD_REQUEST).send({ message: e.message });
-      }
-      if (e.name === "DocumentNotFoundError") {
-        return res.status(NOT_FOUND).send({ message: e.message });
       }
       return res
         .status(DEFAULT)
         .send({ message: "An error has occurred on the server." });
     });
 };
+
+// const { itemId } = req.params;
+// ClothingItem.findByIdAndDelete(itemId)
+//   .orFail()
+//   .then((item) => {
+//     res.status(OKAY_STATUS).send({ data: item });
+//   })
+//   .catch((e) => {
+//     if (e.name === "CastError") {
+//       return res.status(BAD_REQUEST).send({ message: e.message });
+//     }
+//     if (e.name === "DocumentNotFoundError") {
+//       return res.status(NOT_FOUND).send({ message: e.message });
+//     }
+//     return res
+//       .status(DEFAULT)
+//       .send({ message: "An error has occurred on the server." });
+//   });
+// };
 
 const unlikeItem = (req, res) => {
   const { itemId } = req.params;
